server/requirements.txt
asgiref==3.7.2
dj-database-url==2.0.0
Django==4.2.1
djangorestframework==3.14.0
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.0
